fix: refactored util methods to ShapeUtil
